# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alliance University Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, enrollmentNo, rollNo, program, department, semester, email, phone, fatherName, motherName]
      properties:
        id:
          type: integer
        name:
          type: string
        enrollmentNo:
          type: string
        rollNo:
          type: string
        program:
          type: string
        department:
          type: string
        semester:
          type: integer
        section:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        cgpa:
          type: ["number", "null"]
        academicYear:
          type: ["string", "null"]
        admissionNo:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, enrollmentNo, rollNo, program, department, semester, email, phone, fatherName, motherName]
      properties:
        name:
          type: string
        enrollmentNo:
          type: string
        rollNo:
          type: string
        program:
          type: string
        department:
          type: string
        semester:
          type: integer
        section:
          type: string
        email:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        dob:
          type: string
        address:
          type: string
        photoUrl:
          type: string
        academicYear:
          type: string
        admissionNo:
          type: string
        password:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        photoUrl:
          type: string
        semester:
          type: integer
        section:
          type: string
        cgpa:
          type: number
...